requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.4
django-allauth==0.44.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.2.0
pipenv==2021.5.29
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==20.4.7
virtualenv-clone==0.5.4
waitress==2.0.0
